time_complexity_analyzer/requirements.txt
# CI and local dev target: Python 3.12 (see .github/workflows/ci-pipeline.yml, .python-version)
Django>=4.2,<6.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.8
pytest>=8.3.3
pytest-django>=4.1
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
scipy>=1.7.0
numpy>=1.20
gunicorn>=22.0
whitenoise>=6.6
tree-sitter>=0.24,<0.26
tree-sitter-java==0.23.5
tree-sitter-cpp==0.23.4